Extract MeasurementTile to component
